modified test cases for metadata
